# docs/swagger.yaml
basePath: /api
definitions:
  model.CreateAchievementRequest:
    properties:
      category:
        type: string
      description:
        type: string
      points:
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - category
    - description
    - points
    - title
    type: object
  model.LoginRequest:
    description: Login request payload
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
...
  model.RegisterRequest:
    description: User registration request payload
    properties:
      email:
        example: user@example.com
        type: string
      full_name:
        example: New User
        type: string
      password:
        example: "123456"
        type: string
      role_id:
        example: f464ceb1-5481-49cf-99f0-d8f2d66f4506
        type: string
      username:
        example: newuser
        type: string
    required:
    - email
    - full_name
    - password
    - role_id
    - username
    type: object
...
  utils.Meta:
    description: Response metadata
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
      status:
        example: success
        type: string
    type: object
  utils.Response:
    description: Standard API response format
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API untuk sistem manajemen prestasi mahasiswa dengan RBAC
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Achievement Management API
  version: "1.0"
paths:
  /v1/achievements:
    get:
      consumes:
      - application/json
      description: Get list of achievements with role-based filtering and pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        enum:
        - draft
        - submitted
        - verified
        - rejected
        in: query
        name: status
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: string
      - default: created_at
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Achievements retrieved successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get All Achievements
      tags:
      - Achievements
    post:
      consumes:
      - application/json
      description: Create a new achievement (Mahasiswa only)
      parameters:
      - description: Achievement data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateAchievementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Achievement created successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create New Achievement
      tags:
      - Achievements
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/utils.Response'
      summary: User Login
      tags:
      - Authentication
...
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: User Registration
      tags:
      - Authentication
...
  /v1/users:
    get:
      consumes:
      - application/json
      description: Get list of all users (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"